test_gen_project/requirements.txt
fastapi
uvicorn
sqlalchemy
alembic
psycopg2-binary
python-dotenv
pydantic[email]
python-jose[cryptography]
passlib[bcrypt]
sqlalchemy-utils
